+++ src/geophires_x_schema_generator/__init__.py
@@ -18,7 +18,6 @@ from geophires_x.AGSEconomics import AGSEconomics
 from geophires_x.AGSWellBores import AGSWellBores
 from geophires_x.CylindricalReservoir import CylindricalReservoir
 from geophires_x.EconomicsAddOns import EconomicsAddOns
-from geophires_x.EconomicsCCUS import EconomicsCCUS
 from geophires_x.GeoPHIRESUtils import json_dumpse
 from geophires_x.Parameter import Parameter
 from geophires_x.SurfacePlantAGS import SurfacePlantAGS
@@ -73,7 +72,6 @@ class GeophiresXSchemaGenerator:
             (AGSEconomics(dummy_model), 'Economics'),
             (SBTEconomics(dummy_model), 'Economics'),
             (SUTRAEconomics(dummy_model), 'Economics'),
-            (EconomicsCCUS(dummy_model), 'Economics'),
             (EconomicsAddOns(dummy_model), 'Economics'),
         ]